refactor initial version for readability
